rename normalize-variable to normalize-expr
