+++ src/clauses/sub-clauses/OrderBy.ts
@@ -7,7 +7,7 @@ import { CypherASTNode } from "../../CypherASTNode";
 import type { CypherEnvironment } from "../../Environment";
 import type { Expr } from "../../types";
 import { compileCypherIfExists } from "../../utils/compile-cypher-if-exists";
-import { normalizeExpr } from "../../utils/normalize-variable";
+import { normalizeExpr } from "../../utils/normalize-expr";
 
 /** @group Clauses */
 export type Order = "ASC" | "DESC";
