+++ src/expressions/functions/list.ts
@@ -7,7 +7,7 @@ import type { CypherEnvironment } from "../../Environment";
 import type { Variable } from "../../references/Variable";
 import type { Expr } from "../../types";
 import { filterTruthy } from "../../utils/filter-truthy";
-import { normalizeExpr } from "../../utils/normalize-variable";
+import { normalizeExpr } from "../../utils/normalize-expr";
 import { CypherFunction } from "./CypherFunctions";
 
 /**
