+++ src/expressions/functions/string.ts
@@ -6,7 +6,7 @@
 import type { CypherEnvironment } from "../../Environment";
 import type { Expr, NormalizationType } from "../../types";
 import { filterTruthy } from "../../utils/filter-truthy";
-import { normalizeExpr } from "../../utils/normalize-variable";
+import { normalizeExpr } from "../../utils/normalize-expr";
 
 import { CypherFunction } from "./CypherFunctions";
